fix(WDate): fix UB and wrong results for negative (BCE) years in setYmd
WDate::setYmd packed the year into an unsigned int field using a
left-shift on a signed int (y << 16). When the year is negative
(BCE dates), this is undefined behavior in C++11/14/17. Even where
the platform produces a deterministic result, the unsigned storage
caused operator<() to return wrong results (a BCE date compared
greater than any CE date) and year() to return a garbage positive
value (e.g., 65535 for year -1).

The API explicitly supports negative years: fromJulianDay() produces
years as far back as -4713 (Julian Day epoch), and setDate() accepts
any year in [date::year::min(), date::year::max()] = [-32768, 32767].

Fix:
- Change ymd_ from unsigned to int so that operator<() performs a
  signed comparison (correct for BCE vs CE ordering) and year()
  returns the right value via arithmetic right-shift.
- In setYmd(), cast y to unsigned before shifting (shifting unsigned
  is always well-defined in C++) then static_cast<int> the result
  back into ymd_, preserving the two's-complement bit pattern.
- Fix isValid() from (ymd_ > 1) to (ymd_ != 0 && ymd_ != 1).
  The old guard relied on valid dates always being > 1, which holds
  for CE years but fails for BCE years whose packed ymd_ is negative.

All downstream methods (dayOfWeek, daysTo, toJulianDay, toTimePoint,
addDays, addMonths, addYears, toString, fromJulianDay, previousWeekday)
are fixed transitively since they all call year().

+++ src/Wt/WDate.C
@@ -124,7 +124,14 @@ void WDate::setDate(int year, int month, int day)
 
 void WDate::setYmd(int y, int m, int d)
 {
-  ymd_ = (y << 16) | ((m & 0xFF) << 8) | (d & 0xFF);
+  // Cast to unsigned before shifting to avoid UB on negative years (C++11/14/17).
+  // The two's-complement bit pattern is then stored in signed int ymd_ so that
+  // operator<() performs a correct signed comparison and year() returns the right
+  // value via arithmetic right-shift for BCE (negative) years.
+  ymd_ = static_cast<int>(
+      (static_cast<unsigned>(y) << 16) |
+      ((static_cast<unsigned>(m) & 0xFF) << 8) |
+       (static_cast<unsigned>(d) & 0xFF));
 }
 
 bool WDate::isLeapYear(int year)
